Update docs, UPGRADING.md and Boost skill for Support namespace moves
