+++ src/Support/CauserResolver.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Spatie\Activitylog;
+namespace Spatie\Activitylog\Support;
 
 use Closure;
 use Illuminate\Auth\AuthManager;
