+++ tests/ActivityLoggerTest.php
@@ -5,9 +5,9 @@ use Illuminate\Database\Eloquent\JsonEncodingException;
 use Illuminate\Support\Collection;
 use Spatie\Activitylog\Exceptions\CouldNotLogActivity;
 use Spatie\Activitylog\Facades\Activity as ActivityFacade;
-use Spatie\Activitylog\CauserResolver;
-use Spatie\Activitylog\LogOptions;
 use Spatie\Activitylog\Models\Activity;
+use Spatie\Activitylog\Support\CauserResolver;
+use Spatie\Activitylog\Support\LogOptions;
 use Spatie\Activitylog\Test\Enums\NonBackedEnum;
 use Spatie\Activitylog\Test\Models\Article;
 use Spatie\Activitylog\Test\Models\User;
